Python final project port 22.
i was asked to create a python tool that asks the user for a website address, send a ping and if the website is available it will scan for all the TCP Ports. after the scan has completed it needs to perform brute-force on port 22 if it was found open. eventually the code will print the correct username and password for logging in, offer ways to connect and send commands