Python3_final_project
this script it asks the user for an HTTP/S Link, Fixes it into an IP address,  And sends ping to the website. If it's available it starts a tcp scan, And if port 22 (SSH) was found open it offers you to attempt brute-force attack for the login credentials. If it found the correct credentials - it prints them, Shows connection options and offer's to send commands to the remote server, Please use it ONLY for legal purposes, it's up to you.