Python_final_project.py Explained
This script asks the user for an HTTP/S link, Fixes it into an IP address, And sends ping to the website. if it was found available - it starts a tcp scan, and if port 22 (SSH) was found open it offers you to attempt brute-force attack for login credentials. If it found the correct credentials - it prints them, shows connection options and offers to send commands to the remote server, Please use it ONLY for legal purposes, it's up to you.