Bash_SubD-Scan.sh Explained
This is an enumeration scrips written in Bash. it is a compilation of tools which scans for sub-domains under a given URL, Creates a folder for each url, a file for each scan tool and puts everything together in a txt file (subs.txt). Later it sorts out the common file and deleting duplicates, Then it starts sending HTTP/S Requests to the found sub-domains, and puts the ones who responde in a seperate file. atlast it searches for any link with information on the programming language of the active sub-domains and put them in a seperate file.